refactor: rename `identity` module to `account` and update related paths
- Updated package names from `modules.identity` to `modules.account` across the codebase.
- Added new fields (`phone`, `website`, `street`) to promoter registration.
- Removed unused `findByIdWithLock` method from `EventRepository` implementation.
- Adjusted Dockerfile to skip tests using `-Dmaven.test.skip=true`.
