+++ src/main/java/com/gresk/modules/account/application/port/in/LoginUseCase.java
@@ -1,7 +1,7 @@
-package com.gresk.modules.identity.application.port.in;
+package com.gresk.modules.account.application.port.in;
 
 import com.gresk.infrastructure.port.AuthToken;
-import com.gresk.modules.identity.application.command.LoginCommand;
+import com.gresk.modules.account.application.command.LoginCommand;
 
 public interface LoginUseCase {
