+++ src/main/java/com/gresk/modules/account/infrastructure/persistence/JpaAccountRepositoryAdapter.java
@@ -1,8 +1,8 @@
-package com.gresk.modules.identity.infrastructure.persistence;
+package com.gresk.modules.account.infrastructure.persistence;
 
-import com.gresk.modules.identity.domain.model.Account;
-import com.gresk.modules.identity.domain.model.AccountId;
-import com.gresk.modules.identity.domain.port.out.AccountRepositoryPort;
+import com.gresk.modules.account.domain.model.Account;
+import com.gresk.modules.account.domain.model.AccountId;
+import com.gresk.modules.account.domain.port.out.AccountRepositoryPort;
 import com.gresk.shared.domain.valueobject.Email;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
