+++ src/main/java/com/gresk/modules/account/infrastructure/web/dto/AuthResponse.java
@@ -1,3 +1,3 @@
-package com.gresk.modules.identity.infrastructure.web.dto;
+package com.gresk.modules.account.infrastructure.web.dto;
 
 public record AuthResponse(String token, long expiresIn) {}
